+++ src/controls/taxonomyPicker/ITaxonomyPicker.ts
@@ -45,6 +45,10 @@ export interface ITaxonomyPickerProps  {
    */
   disabledTermIds?: string[];
   /**
+   * Specify if you want to disable the child terms when their parent is disabled
+   */
+  disableChildrenOfDisabledParents?: boolean;
+  /**
    * Whether the property pane field is enabled or not.
    */
   disabled?: boolean;
@@ -85,6 +89,7 @@ export interface ITermChanges {
   changedCallback: (term: ITerm, checked: boolean) => void;
   activeNodes?: IPickerTerms;
   disabledTermIds?: string[];
+  disableChildrenOfDisabledParents?: boolean;
 }
 
 
@@ -108,6 +113,7 @@ export interface ITermProps extends ITermChanges {
   termset: string;
   term: ITerm;
   multiSelection: boolean;
+  disabled: boolean;
 }
 
 export interface ITermState {