+++ src/controls/taxonomyPicker/TermPicker.tsx
@@ -8,6 +8,7 @@ import { IWebPartContext } from '@microsoft/sp-webpart-base';
 import * as strings from 'ControlStrings';
 import { Icon } from 'office-ui-fabric-react';
 import { ApplicationCustomizerContext } from '@microsoft/sp-application-base';
+import { ITermSet } from '../../services/ISPTermStorePickerService';
 
 export class TermBasePicker extends BasePicker<IPickerTerm, IBasePickerProps<IPickerTerm>>
 {
@@ -26,11 +27,13 @@ export interface ITermPickerProps {
   allowMultipleSelections : boolean;
   isTermSetSelectable?: boolean;
   disabledTermIds?: string[];
+  disableChildrenOfDisabledParents?: boolean;
 
   onChanged: (items: IPickerTerm[]) => void;
 }
 
 export default class TermPicker extends React.Component<ITermPickerProps, ITermPickerState> {
+  private allTerms: ITermSet = null;
 
   /**
    * Constructor method
@@ -126,14 +129,45 @@ export default class TermPicker extends React.Component<ITermPickerProps, ITermP
       }
       // Filter out the terms which are already set
       const filteredTerms = [];
+      const { disabledTermIds, disableChildrenOfDisabledParents } = this.props;
       for (const term of terms) {
+        let canBePicked = true;
+
         // Check if term is not disabled
-        if (this.props.disabledTermIds && this.props.disabledTermIds.length > 0 && this.props.disabledTermIds.indexOf(term.key) !== -1) {
-          break;
+        if (disabledTermIds && disabledTermIds.length > 0) {
+          // Check if current term need to be disabled
+          if (disabledTermIds.indexOf(term.key) !== -1) {
+            canBePicked = false;
+          } else {
+            // Check if child terms need to be disabled
+            if (disableChildrenOfDisabledParents) {
+              // Check if terms were already retrieved
+              if (!this.allTerms) {
+                this.allTerms = await termsService.getAllTerms(this.props.termPickerHostProps.termsetNameOrID);
+              }
+
+              // Check if there are terms retrieved
+              if (this.allTerms.Terms && this.allTerms.Terms.length > 0) {
+                // Find the disabled parents
+                const disabledParents = this.allTerms.Terms.filter(t => disabledTermIds.indexOf(t.Id) !== -1);
+                // Check if disabled parents were found
+                if (disabledParents && disabledParents.length > 0) {
+                  // Check if the current term lives underneath a disabled parent
+                  const findTerm = disabledParents.filter(pt => term.path.indexOf(pt.PathOfTerm) !== -1);
+                  if (findTerm && findTerm.length > 0) {
+                    canBePicked = false;
+                  }
+                }
+              }
+            }
+          }
         }
-        // Only retrieve the terms which are not yet tagged
-        if (tagList.filter(tag => tag.key === term.key).length === 0) {
-          filteredTerms.push(term);
+
+        if (canBePicked) {
+          // Only retrieve the terms which are not yet tagged
+          if (tagList.filter(tag => tag.key === term.key).length === 0) {
+            filteredTerms.push(term);
+          }
         }
       }
       return filteredTerms;